Working UART in test_uart.c
Got UART working (per testing with XBee radio). Communication to and from RaspberryPi and X-CTU software on PC. However, there seems to be a problem where the baud rate is not actually correctly set until the second run of test_uart.x... probably doing something out of order such that the value is not set until UART is enabled a second time.