UART on RPi and Linux PC
Wrote a program for a Linux PC to interface with XBee over USB. /dev/ttyUSB0

Discovered that Linux (including Raspbian) already has a super-simple interface to communicate using UART, ttyAMA0... so all the low-level stuff in the last commit is sort of useless. Moved the low-level code to drivers/learning/uart_bare.c.

drivers/uart.c now uses the high-level interface to UART, and started writing a QueueBuffer data structure to store data as it comes in. Still need to implement qb_getSize() and do testing.